requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.132.0
google-auth-httplib2==0.2.0
google-auth==2.30.0